+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- attributes to map field to another type
 - wrapper type `egui_struct::Combobox<T>(T)` that through `config: ConfigType` takes list of possible values
 - new config variants for numerics (`SliderStep(..)`, `Combobox(list)`)
 - attribute `on_change_struct` similar to `on_change` but takes expr (may use &mut self (whole struct), not only single field reference)